// nodes/help/type/enums.ts
export declare const enum MessageType {
	Image = 'image',
	File = 'file',
	RichText = 'post',
	Audio = 'audio',
	Video = 'media',
	Card = 'interactive',
	Location = 'location',
	Todo = 'todo',
	CalendarEvent = 'share_calendar_event',
	Text = 'text',
}
// ...
export declare const enum FileType {
	Doc = 'doc',
	Sheet = 'sheet',
	Mindnote = 'mindnote',
	Bitable = 'bitable',
	File = 'file',
	Docx = 'docx',
	Folder = 'folder',
	Shortcut = 'shortcut',
	Slides = 'slides',
}